Consolidate saaj-testsuite tests into BodyTests, ElementTests, HeaderTests
