+++ testing/saaj-testsuite/src/main/java/org/apache/axiom/ts/saaj/SAAJTestSuite.java
@@ -30,15 +30,9 @@ import org.apache.axiom.testutils.suite.Binding;
 import org.apache.axiom.testutils.suite.FanOutNode;
 import org.apache.axiom.testutils.suite.InjectorNode;
 import org.apache.axiom.testutils.suite.LabelBinding;
-import org.apache.axiom.testutils.suite.MatrixTest;
+import org.apache.axiom.testutils.suite.MatrixTestContainer;
 import org.apache.axiom.testutils.suite.MatrixTestNode;
 import org.apache.axiom.testutils.suite.ParentNode;
-import org.apache.axiom.ts.saaj.body.TestAddChildElementReification;
-import org.apache.axiom.ts.saaj.element.TestAddChildElementLocalName;
-import org.apache.axiom.ts.saaj.element.TestAddChildElementLocalNamePrefixAndURI;
-import org.apache.axiom.ts.saaj.element.TestGetOwnerDocument;
-import org.apache.axiom.ts.saaj.element.TestSetParentElement;
-import org.apache.axiom.ts.saaj.header.TestExamineMustUnderstandHeaderElements;
 import org.apache.axiom.ts.soap.SOAPSpec;
 
 public class SAAJTestSuite {
@@ -66,11 +60,8 @@ public class SAAJTestSuite {
                                     }
                                 },
                                 new ParentNode(
-                                        new MatrixTest(TestAddChildElementReification.class),
-                                        new MatrixTest(TestExamineMustUnderstandHeaderElements.class),
-                                        new MatrixTest(TestAddChildElementLocalName.class),
-                                        new MatrixTest(TestAddChildElementLocalNamePrefixAndURI.class),
-                                        new MatrixTest(TestSetParentElement.class),
-                                        new MatrixTest(TestGetOwnerDocument.class)))));
+                                        new MatrixTestContainer(BodyTests.class),
+                                        new MatrixTestContainer(ElementTests.class),
+                                        new MatrixTestContainer(HeaderTests.class)))));
     }
 }
