--- testing/saaj-testsuite/src/main/java/org/apache/axiom/ts/saaj/element/TestAddChildElementLocalName.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.axiom.ts.saaj.element;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.google.inject.Inject;
-import jakarta.xml.soap.SOAPElement;
-import jakarta.xml.soap.SOAPFactory;
-import org.junit.jupiter.api.function.Executable;
-
-/** Tests the behavior of {@link SOAPElement#addChildElement(String)}. */
-public class TestAddChildElementLocalName implements Executable {
-    @Inject
-    private SOAPFactory soapFactory;
-
-    @Override
-    public void execute() throws Throwable {
-        SOAPElement root = soapFactory.createElement("root", "p", "urn:test");
-        SOAPElement element = root.addChildElement("child");
-        assertThat(element.getLocalName()).isEqualTo("child");
-        assertThat(element.getNamespaceURI()).isNull();
-        assertThat(element.getPrefix()).isNull();
-        assertThat(element.getParentNode()).isSameAs(root);
-        assertThat(element.getAttributes().getLength()).isEqualTo(0);
-    }
-}
